Refactor the SG global so it can be called from multiple processes
